front/requirements.txt
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
decorator
exceptiongroup
fastcore
fasteners
fastjsonschema
filelock
gitdb
GitPython
h11
idna
importlib-metadata
Jinja2
json5
jsonpointer
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
numpy
packaging
pandas
Pillow
protobuf
pyarrow
pydeck
Pygments
Pympler
python-dateutil
python-multipart
pytz
pytz-deprecation-shim
referencing
requests
rich
rpds-py
six
smmap
sniffio
starlette
streamlit
sympy
tenacity
toml
toolz
typing_extensions
tzdata
tzlocal
urllib3
validators
zipp
